pyproject.toml
[tool.poetry]
name = "python-kafka-example"
version = "0.1.0"
description = ""
authors = ["Manuel Castillo <manucalop@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
confluent-kafka = "^1.9.2"
pydantic = "^1.9.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
